Fix typescript error on build
Address the error

```
Type error: Property 'id' has no initializer and is not definitely
assigned in the constructor.
```

The way that TypeORM uses decorators to determine field types but then
also uses a builder pattern to create and persist objects is kind of
awkward with TypeScript. We can work around it by not using the entity
constructor directly and instead using a wrapper function that returns a
more accurate type.